+++ CHANGELOG.md
@@ -43,6 +43,7 @@ Core Updates:
 * Export button has contextual text for images / data
 * Fixed a bug where user config files were loaded twice
 * Refactored code so that the order of modules can be changed in the user config
+* Beefed up code + docs in scatter plots back end and multiple bar plots.
 
 #### [v0.8](https://github.com/ewels/MultiQC/releases/tag/v0.8) - 2016-09-26
 Module updates: